requirements.txt
Flask==2.2.3
requests==2.28.2
Werkzeug==2.2.3
mysql-connector-python==8.0.32
beautifulsoup4==4.11.2
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
idna==3.4
itsdangerous==2.1.2
Jninja2==3.1.2
MarkupSafe==2.1.2
protobuf==3.20.3
soupsieve==2.4
urllib3==1.26.14

